daemon: last_activity_epoch_ms crosses the wire as u64, not u128
Measured 2026-09-02 14:0x, first restart onto the fs-truth build: every
`server snapshot` CLIENT parse failed with `u128 is not supported` —
serde_json without `arbitrary_precision` refuses to deserialize u128, and
the field is a WIRE field. Epoch ms fits u64 for the next 500 millennia;
the u128 was borrowed from StartpageDurableRow's internal field and never
belonged on the wire. Daemon fill, both verbs' folds, and the shell's
session conversion widened back to u128 only at the internal row boundary.

+++ crates/yggterm-server/src/daemon.rs
@@ -6696,7 +6696,7 @@ impl DaemonRuntime {
             session.last_activity_epoch_ms = self
                 .terminals
                 .session_idle_for_ms(&runtime_path)
-                .map(|idle_ms| (now_ms.saturating_sub(idle_ms)) as u128);
+                .map(|idle_ms| now_ms.saturating_sub(idle_ms));
             // ⛔ ONLY where this daemon has no answer of its own. A scrape of a
             // screen we own is the freshest truth there is; a proxied flag may
             // be up to one refresh old, so it fills a hole and never overwrites.
